Moved functions from ‘javalinker’ to ‘java_module’
